# docs/swagger.yaml
basePath: /v1/api
definitions:
  CONVERDA_internal_apps_controller_dto.APIKeyFullResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      name:
        type: string
      plain_key:
        description: Only returned once on creation/rotation
        type: string
      prefix:
        type: string
      revoked_at:
        type: string
      suffix:
        type: string
    type: object
  CONVERDA_internal_apps_controller_dto.APIKeyResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      name:
        type: string
      prefix:
        type: string
      revoked_at:
        type: string
      suffix:
        type: string
    type: object
  CONVERDA_internal_apps_controller_dto.AppResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      environments:
        items:
          $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.EnvironmentResponse'
        type: array
      id:
        type: string
      name:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
  CONVERDA_internal_apps_controller_dto.CreateAppRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  CONVERDA_internal_apps_controller_dto.CreateEnvironmentRequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
...
  CONVERDA_internal_apps_controller_dto.EnvironmentResponse:
    properties:
      app_id:
        type: string
      created_at:
        type: string
      environment_code:
        type: string
      id:
        type: string
      keys:
        items:
          $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.APIKeyResponse'
        type: array
    type: object
...
  CONVERDA_internal_apps_controller_dto.RotateKeyRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  CONVERDA_internal_apps_controller_dto.SystemEnvironmentResponse:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
  CONVERDA_internal_apps_controller_dto.UpdateAppRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a server for a Go Drunk Backend API, demonstrating DDD principles.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Drunk Backend API by DDD
  version: "1.0"
paths:
...
  /api-keys/{key_id}/revoke:
    post:
      parameters:
      - description: Key ID
        in: path
        name: key_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Revoke API Key
      tags:
      - Apps
  /apps/{app_id}:
    delete:
      description: Delete an app
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete app
      tags:
      - Apps
    get:
      description: Get details of a specific app
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.AppResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get app
      tags:
      - Apps
    put:
      consumes:
      - application/json
      description: Update an existing app
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.UpdateAppRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.AppResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update app
      tags:
      - Apps
  /apps/{app_id}/environments:
    get:
      description: List all environments for an app
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.EnvironmentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List environments
      tags:
      - Apps
    post:
      consumes:
      - application/json
      description: Create a new environment for an app
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: string
      - description: Environment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.CreateEnvironmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.EnvironmentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create environment
      tags:
      - Apps
...
  /environments/{id}/api-keys:
    get:
      parameters:
      - description: Env ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.APIKeyResponse'
            type: array
      security:
      - BearerAuth: []
      summary: List API Keys
      tags:
      - Apps
  /environments/{id}/api-keys/rotate:
    post:
      consumes:
      - application/json
      description: Revokes all old keys and generates a new one
      parameters:
      - description: Env ID
        in: path
        name: id
        required: true
        type: string
      - description: Key data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.RotateKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.APIKeyFullResponse'
      security:
      - BearerAuth: []
      summary: Rotate API Key
      tags:
      - Apps
...
  /system/environments:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.SystemEnvironmentResponse'
            type: array
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: List system environments
      tags:
      - System
...
  /tenants/{id}/apps:
    get:
      description: List all apps for a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.AppResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List apps
      tags:
      - Apps
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: App data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.CreateAppRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.AppResponse'
      security:
      - BearerAuth: []
      summary: Create a new app
      tags:
      - Apps
...
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"